+++ crates/config-loader/src/toolchains_config_ext.rs
@@ -137,7 +137,7 @@ impl ToolchainsConfigExt {
                 }
                 #[cfg(debug_assertions)]
                 "tc-tier1" | "tc-tier2" | "tc-tier2-reqs" | "tc-tier2-setup-env" | "tc-tier3"
-                | "tc-tier3-reqs" => {
+                | "tc-tier3-reqs" | "tc-tier3-tool" => {
                     use proto_core::{UnresolvedVersionSpec, warpgate::find_debug_locator};
 
                     config.plugin = Some(